I changed the emodex set_heart func to a float
I did this because the user's heart is in integers but the robot's heart should be in float values, as any perceived emotional intent from the texts it receives effects it in a 2:5 ration average 2 of the emotional intent, 3 of the current emotion, added together, divided by 5. The robot starts out as the user's initial emotions, then. as each of them talk the emotional intent the robot expresses and the emotional intent the human expresses change the robot's emotional state.

+++ Emodex.py
@@ -154,7 +154,7 @@ my_other = {
 
 def set_heart(my_heart, key, subkey, value):
     if key in my_heart and subkey in my_heart[key]:
-        if isinstance(value, int) and 0 <= value <= 5:
+        if isinstance(value, float) and 0 <= value <= 5:
             my_heart[key][subkey] = value
         elif key == "focus" and subkey in my_heart[key] and isinstance(value, int) and 0 <= value <= 1:
             my_heart[key][subkey] = value
